requirements.txt
bottle==0.12.9
RPi.GPIO==0.5.11
wheel==0.24.0
